Tons of changes
While I was trying at first to incorporate ts in my code I just didn't understand the compiler and how it worked it was hindering my ability to actually make something so in this one I have removed all my ts code that I had tried to do and now I'll be doing just vanilla js for this one at least. The next one I'll be just implementing a media player. Also in this one I moved all my js files in a script folder just because it looks better