Added a login page and a signup page
Now I have added login and signup pages, They both redirect when if you have valid credentials but I still have to implement the sqlquesries that would actualy make sure that the credentials are valid but the functionality is there,

the signup page also has is a good password checker and a confirm password thing it has very nice response the only thing it lacks right now is separating the good chars from the numbers and performing separate checks

The way I have achieved redirecting is a bit weird I think, I just removed the DOM and added a new DOM basically removing the html element and writing the other html file to the dom. so Im going to fix that later

I have also made the media player a lot better it doesn't work as of now because of a small recursive bug but I just got bored and decided to do something else,  after I do the sql queries I am going to try and do media player.

There was no speech systhesis variable in the text to speech so I added that

Right Now, The only thing I am confused about is that how am I going to give the information about the user to the client after it has been redirected, I think I have a solution now because of the way that I "redirected" it, all the prev js files are still in the dom so what I can still do is just just make that user object equal to the user object I have in the user.js or the other way to do it is I guess some kind of map object that I could also do for sure